[Maps] convert redux actions to TS (#67218)
* [Maps] convert redux actions to TS

* move ts-ignore comment to right location

* eslint
