+++ x-pack/plugins/maps/public/actions/ui_actions.ts
@@ -4,9 +4,12 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
+import { Dispatch } from 'redux';
+import { MapStoreState } from '../reducers/store';
 import { getFlyoutDisplay } from '../selectors/ui_selectors';
-import { FLYOUT_STATE } from '../reducers/ui';
-import { setSelectedLayer, trackMapSettings } from './map_actions';
+import { FLYOUT_STATE, INDEXING_STAGE } from '../reducers/ui';
+import { trackMapSettings } from './map_actions';
+import { setSelectedLayer } from './layer_actions';
 
 export const UPDATE_FLYOUT = 'UPDATE_FLYOUT';
 export const CLOSE_SET_VIEW = 'CLOSE_SET_VIEW';
@@ -26,19 +29,19 @@ export function exitFullScreen() {
   };
 }
 
-export function updateFlyout(display) {
+export function updateFlyout(display: FLYOUT_STATE) {
   return {
     type: UPDATE_FLYOUT,
     display,
   };
 }
 export function openMapSettings() {
-  return (dispatch, getState) => {
+  return (dispatch: Dispatch, getState: () => MapStoreState) => {
     const flyoutDisplay = getFlyoutDisplay(getState());
     if (flyoutDisplay === FLYOUT_STATE.MAP_SETTINGS_PANEL) {
       return;
     }
-    dispatch(setSelectedLayer(null));
+    dispatch<any>(setSelectedLayer(null));
     dispatch(trackMapSettings());
     dispatch(updateFlyout(FLYOUT_STATE.MAP_SETTINGS_PANEL));
   };
@@ -53,7 +56,7 @@ export function openSetView() {
     type: OPEN_SET_VIEW,
   };
 }
-export function setIsLayerTOCOpen(isLayerTOCOpen) {
+export function setIsLayerTOCOpen(isLayerTOCOpen: boolean) {
   return {
     type: SET_IS_LAYER_TOC_OPEN,
     isLayerTOCOpen,
@@ -65,35 +68,35 @@ export function enableFullScreen() {
     isFullScreen: true,
   };
 }
-export function setReadOnly(isReadOnly) {
+export function setReadOnly(isReadOnly: boolean) {
   return {
     type: SET_READ_ONLY,
     isReadOnly,
   };
 }
 
-export function setOpenTOCDetails(layerIds) {
+export function setOpenTOCDetails(layerIds?: string[]) {
   return {
     type: SET_OPEN_TOC_DETAILS,
     layerIds,
   };
 }
 
-export function showTOCDetails(layerId) {
+export function showTOCDetails(layerId: string) {
   return {
     type: SHOW_TOC_DETAILS,
     layerId,
   };
 }
 
-export function hideTOCDetails(layerId) {
+export function hideTOCDetails(layerId: string) {
   return {
     type: HIDE_TOC_DETAILS,
     layerId,
   };
 }
 
-export function updateIndexingStage(stage) {
+export function updateIndexingStage(stage: INDEXING_STAGE | null) {
   return {
     type: UPDATE_INDEXING_STAGE,
     stage,
