+++ x-pack/plugins/maps/public/angular/get_initial_layers.d.ts
@@ -4,4 +4,9 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
-export function getInitialLayers(layerListJSON?: string, initialLayers?: unknown[]): unknown[];
+import { LayerDescriptor } from '../../common/descriptor_types';
+
+export function getInitialLayers(
+  layerListJSON?: string,
+  initialLayers?: LayerDescriptor[]
+): LayerDescriptor[];
