+++ x-pack/plugins/maps/public/connected_components/widget_overlay/layer_control/layer_toc/toc_entry/toc_entry_actions_popover/index.ts
@@ -28,16 +28,16 @@ function mapStateToProps(state: MapStoreState) {
 function mapDispatchToProps(dispatch: Dispatch<AnyAction>) {
   return {
     cloneLayer: (layerId: string) => {
-      dispatch(cloneLayer(layerId));
+      dispatch<any>(cloneLayer(layerId));
     },
     fitToBounds: (layerId: string) => {
-      dispatch(fitToLayerExtent(layerId));
+      dispatch<any>(fitToLayerExtent(layerId));
     },
     removeLayer: (layerId: string) => {
-      dispatch(removeLayer(layerId));
+      dispatch<any>(removeLayer(layerId));
     },
     toggleVisible: (layerId: string) => {
-      dispatch(toggleLayerVisible(layerId));
+      dispatch<any>(toggleLayerVisible(layerId));
     },
   };
 }
