+++ x-pack/plugins/maps/public/embeddable/map_embeddable.tsx
@@ -55,6 +55,7 @@ import { getMapCenter, getMapZoom, getHiddenLayerIds } from '../selectors/map_se
 import { MAP_SAVED_OBJECT_TYPE } from '../../common/constants';
 import { RenderToolTipContent } from '../classes/tooltips/tooltip_property';
 import { getUiActions, getCoreI18n } from '../kibana_services';
+import { LayerDescriptor } from '../../common/descriptor_types';
 
 import { MapEmbeddableInput, MapEmbeddableConfig } from './types';
 export { MapEmbeddableInput, MapEmbeddableConfig };
@@ -69,7 +70,7 @@ export class MapEmbeddable extends Embeddable<MapEmbeddableInput, MapEmbeddableO
 
   private _renderTooltipContent?: RenderToolTipContent;
   private _eventHandlers?: EventHandlers;
-  private _layerList: unknown[];
+  private _layerList: LayerDescriptor[];
   private _store: MapStore;
   private _subscription: Subscription;
   private _prevTimeRange?: TimeRange;
@@ -147,7 +148,7 @@ export class MapEmbeddable extends Embeddable<MapEmbeddableInput, MapEmbeddableO
     this._prevTimeRange = timeRange;
     this._prevQuery = query;
     this._prevFilters = filters;
-    this._store.dispatch(
+    this._store.dispatch<any>(
       setQuery({
         filters: filters.filter((filter) => !filter.meta.disabled),
         query,
@@ -218,9 +219,9 @@ export class MapEmbeddable extends Embeddable<MapEmbeddableInput, MapEmbeddableO
       );
     }
 
-    this._store.dispatch(replaceLayerList(this._layerList));
+    this._store.dispatch<any>(replaceLayerList(this._layerList));
     if (this.input.hiddenLayers) {
-      this._store.dispatch(setHiddenLayers(this.input.hiddenLayers));
+      this._store.dispatch<any>(setHiddenLayers(this.input.hiddenLayers));
     }
     this._dispatchSetQuery(this.input);
     this._dispatchSetRefreshConfig(this.input);
@@ -248,9 +249,9 @@ export class MapEmbeddable extends Embeddable<MapEmbeddableInput, MapEmbeddableO
     });
   }
 
-  async setLayerList(layerList: unknown[]) {
+  async setLayerList(layerList: LayerDescriptor[]) {
     this._layerList = layerList;
-    return await this._store.dispatch(replaceLayerList(this._layerList));
+    return await this._store.dispatch<any>(replaceLayerList(this._layerList));
   }
 
   addFilters = (filters: Filter[]) => {
