+++ x-pack/plugins/maps/public/embeddable/types.ts
@@ -9,14 +9,14 @@ import { MapSettings } from '../reducers/map';
 // eslint-disable-next-line @kbn/eslint/no-restricted-paths
 import { EmbeddableInput } from '../../../../../src/plugins/embeddable/public/lib/embeddables';
 import { Filter, Query, RefreshInterval, TimeRange } from '../../../../../src/plugins/data/common';
-import { MapCenterAndZoom } from '../../common/descriptor_types';
+import { LayerDescriptor, MapCenterAndZoom } from '../../common/descriptor_types';
 
 export interface MapEmbeddableConfig {
   editUrl?: string;
   indexPatterns: IIndexPattern[];
   editable: boolean;
   title?: string;
-  layerList: unknown[];
+  layerList: LayerDescriptor[];
   settings?: MapSettings;
 }
