+++ x-pack/plugins/siem/public/network/components/embeddables/embedded_map_helpers.tsx
@@ -109,6 +109,7 @@ export const createEmbeddable = async (
 
   if (!isErrorEmbeddable(embeddableObject)) {
     embeddableObject.setRenderTooltipContent(renderTooltipContent);
+    // @ts-ignore
     await embeddableObject.setLayerList(getLayerList(indexPatterns));
   }
